+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## [Unreleased]
+## `2.0.0` (2018-04-22)
+
+- Use MarkDownDOM to parse the unordered checkbox list blocks
+- Use `lineCount` and `lineAt` instead of my own regex to be 100 % compatible with VS Code behavior
+
+## `1.0.0` (2018-04-11)
 
 - Initial release